added live preview, made forms secure, added deletion of filters, working on multiple filters at once

+++ utils/upload_webcam.php
@@ -4,6 +4,11 @@ use \ablin42\database;
 if (isset($_POST['submit_cam']) && !empty($_POST['img_url']) && !empty($_POST['id_user_cam'])
     && !empty($_POST['img_name_cam']) && !empty($_POST['filter']) && !empty($_POST['tmp_img']))
 {
+    if (!check_length($_POST['img_name_cam'], 1, 64))
+    {
+        echo alert_bootstrap("warning", "Your <b>title</b> has to be 1 character minimum and 64 characters maximum!", "text-align: center;");
+        return ;
+    }
     unlink("{$_POST['tmp_img']}");
     $db = database::getInstance('camagru');
     $req = $db->query( "SELECT MAX(id) as last_id FROM `image`");
@@ -19,7 +24,7 @@ if (isset($_POST['submit_cam']) && !empty($_POST['img_url']) && !empty($_POST['i
     imagealphablending($photo, true);
     imagesavealpha($filter, true);
 
-    $info = get_filter_position($_POST['filter']);
+    $info = get_filter_position(htmlspecialchars(trim($_POST['filter'])));
 
     imagecopy($photo, $filter, $info['dst_x'], $info['dst_y'], 0, 0, $info['src_w'], $info['src_h']);
     imagepng($photo, $path);
@@ -27,9 +32,9 @@ if (isset($_POST['submit_cam']) && !empty($_POST['img_url']) && !empty($_POST['i
     imagedestroy($filter);
 
     $path = "/Camagru/images/{$id_img}.png";
-    $attributes['id_user'] = $_POST['id_user_cam'];
+    $attributes['id_user'] = htmlspecialchars(trim($_POST['id_user_cam']));
     $attributes['path'] = $path;
-    $attributes['name'] = $_POST['img_name_cam'];
+    $attributes['name'] =  htmlspecialchars(trim($_POST['img_name_cam']));
     $req = $db->prepare("INSERT INTO `image` (`id_user`, `path`, `name`, `date`) VALUES (:id_user, :path, :name, NOW())", $attributes);
     echo alert_bootstrap("success", "<b>Congratulations!</b> Your picture has been posted!", "text-align: center;");
 }
